moved AddPaneModal.js, Pane.js, VerseDisplay.js, View.js components to the components folder
